+++ trane/core/prediction_problem_generator.py
@@ -68,7 +68,11 @@ class PredictionProblemGenerator:
             assert isinstance(col, str)
             assert isinstance(column_schema, ColumnSchema)
             assert col in self.df.columns
-            assert column_schema.logical_type.dtype == str(self.df[col].dtype)
+            assert column_schema.logical_type.dtype == self.df[col].dtype
+        for col in self.df.columns:
+            assert col in self.table_meta.keys()
+
+        assert len(self.df.columns) == len(self.table_meta.keys())
 
         if self.entity_col:
             entity_col_type = self.table_meta[self.entity_col]