+++ docs/changelog.md
@@ -6,6 +6,7 @@ Changelog
 v0.6.0 (, 2023)
 ===============
 * Enhancements
+    * Add pyarrow dependency and use pyarrow backed dtypes [#120][#120]
 * Fixes
     * Rename `_execute_operations_on_df` to `target` in executed prediction problem dataframe [#124][#124]
     * Clean up operation description generation [#118][#118]
@@ -14,6 +15,7 @@ v0.6.0 (, 2023)
 
     [#124]: <https://github.com/trane-dev/Trane/pull/124>
     [#118]: <https://github.com/trane-dev/Trane/pull/118>
+    [#120]: <https://github.com/trane-dev/Trane/pull/120>
 
 
 v0.5.0 (July 27, 2023)